requirements.txt
# Core
anthropic>=0.40.0
openai>=1.0.0

# Training
torch>=2.1.0
transformers>=4.45.0
peft>=0.12.0
trl>=0.11.0
unsloth>=2024.11.0
bitsandbytes>=0.43.0
accelerate>=0.34.0
datasets>=2.20.0

# Embedding & Retrieval
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# Graph
networkx>=3.0

# Data
jsonlines>=4.0.0
tqdm>=4.66.0
pyyaml>=6.0

# Eval
numpy>=1.26.0
scikit-learn>=1.4.0

# Cloud
google-cloud-bigquery>=3.13.0
google-cloud-storage>=2.14.0
db-dtypes>=1.2.0

# Utils
python-dotenv>=1.0.0
